Release 0.9.8

+++ Lib/BootStrap.ps1
@@ -36,8 +36,8 @@ if (Test-Path -Path "$env:SystemDrive\BootStrap\localhost.mof") {
     } #end while
 } #end if localhost.mof
 } #end CoreCLR bootstrap scriptblock
-        
-        $sciptBlock = {
+
+        $scriptBlock = {
 ## Lability DSC Bootstrap
 $VerbosePreference = 'Continue';
 $DebugPreference = 'Continue';
@@ -50,6 +50,9 @@ certutil.exe -addstore -f "Root" "$env:SYSTEMDRIVE\BootStrap\LabRoot.cer";
 
 <#CustomBootStrapInjectionPoint#>
 
+## Account for large configurations being "pushed" and increase the default from 500KB to 1024KB (1MB)
+Set-Item -Path WSMan:\localhost\MaxEnvelopeSizekb -Value 1024 -Force -Verbose;
+
 if (Test-Path -Path "$env:SystemDrive\BootStrap\localhost.meta.mof") {
     Set-DscLocalConfigurationManager -Path "$env:SystemDrive\BootStrap\" -Verbose;
 }
@@ -60,32 +63,42 @@ if (Test-Path -Path $localhostMofPath) {
         ## Convert the .mof to v4 compatible - credit to Mike Robbins
         ## http://mikefrobbins.com/2014/10/30/powershell-desired-state-configuration-error-undefined-property-configurationname/
         $mof = Get-Content -Path $localhostMofPath;
-        $mof -replace '^\sName=.*;$|^\sConfigurationName\s=.*;$' | Set-Content -Path $localhostMofPath -Encoding Ascii -Force;
+        $mof -replace '^\sName=.*;$|^\sConfigurationName\s=.*;$' | Set-Content -Path $localhostMofPath -Encoding Unicode -Force;
     }
     while ($true) {
         ## Replay the configuration until the LCM bloody-well takes it!
         try {
-            Start-DscConfiguration -Path "$env:SystemDrive\Bootstrap\" -Force -Wait -Verbose -ErrorAction Stop;
-            break;
+            if (Test-Path -Path "$env:SystemRoot\System32\Configuration\Pending.mof") {
+                Start-DscConfiguration -UseExisting -Force -Wait -Verbose -ErrorAction Stop;
+                break;
+            }
+            else {
+                Start-DscConfiguration -Path "$env:SystemDrive\Bootstrap\" -Force -Wait -Verbose -ErrorAction Stop;
+                break;
+            }
         }
         catch {
             Write-Error -Message $_;
-            ## SIGH. Try removing the configuration and restarting WMI..
-            Remove-DscConfigurationDocument -Stage Current,Pending,Previous -Force;
-            Restart-Service -Name Winmgmt -Force;
+            ## SIGH. Try restarting WMI..
+            if (-not ($interation % 10)) {
+                ## SIGH. Try removing the configuration and restarting WMI..
+                Remove-DscConfigurationDocument -Stage Current,Pending,Previous -Force;
+                Restart-Service -Name Winmgmt -Force;
+            }
             Start-Sleep -Seconds 5;
+            $interation++;
         }
     } #end while
 } #end if localhost.mof
 
 Stop-Transcript;
 } #end bootstrap scriptblock
-        
+
         if ($CoreCLR) {
             return $coreCLRScriptBlock;
         }
         else {
-            return $sciptBlock;
+            return $scriptBlock;
         }
     } #end process
 } #end function NewBootStrap
@@ -134,7 +147,7 @@ function SetBootStrap {
         ## Destination Bootstrap directory path.
         [Parameter(Mandatory, ValueFromPipeline)]
         [System.String] $Path,
-        
+
         ## Custom bootstrap script
         [Parameter(ValueFromPipelineByPropertyName)] [ValidateNotNullOrEmpty()]
         [System.String] $CustomBootStrap,
@@ -166,7 +179,7 @@ function ResolveCustomBootStrap {
         [Parameter(Mandatory, ValueFromPipeline)]
         [ValidateSet('ConfigurationFirst','ConfigurationOnly','Disabled','MediaFirst','MediaOnly')]
         [System.String] $CustomBootstrapOrder,
-        
+
         ## Node/configuration custom bootstrap script
         [Parameter(ValueFromPipelineByPropertyName)] [AllowNull()]
         [System.String] $ConfigurationCustomBootStrap,
